requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
redis==5.0.1
python-dotenv==1.0.0